Removed FASTLed as incompatible with ESP IDF 5.x
Moved BusI2C to RaftI2C
Added separate classes for AS5600 and MT6701

+++ components/Scader/ScaderDoors/DoorStrike.cpp
@@ -210,7 +210,7 @@ String DoorStrike::getStatusJson(bool incBraces) const
                         incBraces ? "{" : "", 
                         getDoorLockedStr(lockedEnum), 
                         getDoorOpenSenseStr(openEnum),
-                        timeBeforeLockMs,
+                        (int)timeBeforeLockMs,
                         incBraces ? "}" : "");
     return statusStr;
 }
